+++ RELEASENOTES.md
@@ -19,6 +19,7 @@
 - Fix intermission font character width setting. (fixes Tele-Direct intermission font numbers).
 - Fix parsing both UMAPINFO and ZMAPINFO when present. ZMAPINFO takes priority. (fixes Crematomania MAP30 endgame).
 - Fix chainsaw/punch always using zero pitch when autoaim is on and there nothing to aim at.
+- Fix issue with software emulation that would cause issues with sprites rendering over lowers when a two-sided middle wall was set on the opposite side.
 
 ## Misc:
 - Refactor of old Status Bar renderer to data-driven SBARDEF format.